Moved home page to pages directory
