+++ src/app/pages/home/home.page.html
@@ -1,0 +1,77 @@
+<ion-header [translucent]="true">
+	<ion-toolbar>
+		<ion-title>
+			Home
+		</ion-title>
+	</ion-toolbar>
+</ion-header>
+
+<ion-content [fullscreen]="true">
+	<!-- <ion-slides>
+		<ion-slide> -->
+			<ion-card>
+				<ion-card-header>
+					<ion-card-title>Characters</ion-card-title>
+				</ion-card-header>
+
+				<ion-card-content>
+					<ion-list>
+						<ion-item (click)="PresentModal()">
+							<ion-label>Stór</ion-label>
+						</ion-item>
+						<ion-item (click)="PresentModal()">
+							<ion-label>Orvar</ion-label>
+						</ion-item>
+						<ion-item (click)="PresentModal()">
+							<ion-label>Adelyn</ion-label>
+						</ion-item>
+						<ion-item (click)="PresentModal()">
+							<ion-label>Nyx</ion-label>
+						</ion-item>
+					</ion-list>
+				</ion-card-content>
+			</ion-card>
+
+			<ion-card>
+				<ion-card-header>
+					<ion-card-title>Associates</ion-card-title>
+				</ion-card-header>
+
+				<ion-card-content>
+					<ion-list>
+						<ion-item (click)="PresentModal()">
+							<ion-label>Jarec</ion-label>
+						</ion-item>
+					</ion-list>
+				</ion-card-content>
+			</ion-card>
+
+			<ion-card>
+				<ion-card-header>
+					<ion-card-title>Non-player Characters</ion-card-title>
+				</ion-card-header>
+
+				<ion-card-content>
+					<ion-list>
+						<ion-item (click)="PresentModal()">
+							<ion-label>The Kynfæri Brothers</ion-label>
+						</ion-item>
+					</ion-list>
+				</ion-card-content>
+			</ion-card>
+		<!-- </ion-slide>
+		<ion-slide>
+			<ion-card>
+				<ion-card-header>
+					<ion-card-title>Winds of War</ion-card-title>
+				</ion-card-header>
+
+				<ion-card-content>
+					<iframe name="Winds of War"
+						src="https://www.profounddecisions.co.uk/empire-wiki/383YE_Autumn_Equinox_winds_of_war">
+					</iframe>
+				</ion-card-content>
+			</ion-card>
+		</ion-slide>
+	</ion-slides> -->
+</ion-content>
