+++ src/app/pages/home/home.page.scss
@@ -1,0 +1,3 @@
+ion-content ion-toolbar {
+  --background: translucent;
+}
